Task 7

+++ 0x00-python_variable_annotations/7-to_kv.py
@@ -3,7 +3,8 @@
 
 from typing import Union, Tuple
 
-def to_kv(k: str, v: Union[float, int]) -> Tuple[str, float]:
+
+def to_kv(k: str, v: Union[int, float]) -> Tuple[str, float]:
     """
     Create tuple with k and the square of int/float v.
